requirements.txt
https://github.com/pyrogram/pyrogram/archive/develop.zip
speedtest-cli
